Add forwarding
add backpropagation

+++ includes/upload.h
@@ -14,8 +14,8 @@ public:
     File();
     File(char *newFileName);
     void chargeFileName(char *newFileName);
-    vector< vector< vector<float> > > read();
-    void write(vector< vector< vector<float> > > poids);
+    vector< vector< vector<double> > > read();
+    void write(vector< vector< vector<double> > > poids);
 };
 
 #endif
